Move rate/conc control to IsStream/Transform
