+++ src/Streamly/Internal/Data/Stream/IsStream/Common.hs
@@ -50,7 +50,7 @@ import Control.Monad.IO.Class (MonadIO(..))
 import Foreign.Storable (Storable)
 import Streamly.Internal.Data.Array.Storable.Foreign.Types (Array)
 import Streamly.Internal.Data.Fold.Types (Fold (..))
-import Streamly.Internal.Data.Stream.Combinators (maxYields)
+import Streamly.Internal.Data.Stream.IsStream.Combinators (maxYields)
 import Streamly.Internal.Data.Stream.Prelude (fromStreamS, toStreamS)
 import Streamly.Internal.Data.Stream.Serial (SerialT)
 import Streamly.Internal.Data.Stream.StreamD (fromStreamD, toStreamD)
